Improve detection of possible include guards
We used to consider features to be possible include guards if they were
defined (without a value) conditional to themselves, for example:

 #ifndef TEST_H
 #define TEST_H
 #endif

TEST_H would be disqualified from the regular list of features in this
case. Now it can also be un-disqualified if it's also defined
somewhere else, for example:

 #define AA

 #ifndef AA
 #define AA
 #endif

The list of these identifiers are also stored as "other_defines" in the
results json.